requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.8.0
fsspec==2024.12.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.11.1
json-repair==0.44.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-experimental==0.3.4
langchain-neo4j==0.5.0
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langsmith==0.4.38
marshmallow==3.26.1
multidict==6.7.0
mypy-extensions==1.1.0
neo4j==5.28.2
neo4j-graphrag==1.10.1
numpy==2.3.4
openai==2.6.1
orjson==3.11.4
packaging==25.0
propcache==0.4.1
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-settings==2.11.0
pypdf==6.1.3
python-dotenv==1.2.1
pytz==2025.2
pyyaml==6.0.3
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
scipy==1.16.3
sniffio==1.3.1
sqlalchemy==2.0.44
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
types-pyyaml==6.0.12.20250915
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
urllib3==2.5.0
yarl==1.22.0
zstandard==0.25.0
